requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-environ>=0.11.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
reportlab>=4.0.0
Pillow>=10.0.0

